+++ modules/AWSPowerShell/Cmdlets/IoT/Basic/Add-IOTPrincipalPolicy-Cmdlet.cs
@@ -31,7 +31,8 @@ namespace Amazon.PowerShell.Cmdlets.IOT
     /// Attaches the specified policy to the specified principal (certificate or other credential).
     /// 
     ///  
-    /// <para><b>Note:</b> This action is deprecated. Please use <a>AttachPolicy</a> instead.
+    /// <para><b>Note:</b> This action is deprecated and works as expected for backward compatibility,
+    /// but we won't add enhancements. Use <a>AttachPolicy</a> instead.
     /// </para><para>
     /// Requires permission to access the <a href="https://docs.aws.amazon.com/service-authorization/latest/reference/list_awsiot.html#awsiot-actions-as-permissions">AttachPrincipalPolicy</a>
     /// action.